+++ example/example/settings.py
@@ -53,7 +53,7 @@ ROBOTSTXT_OBEY = True
 # Enable or disable downloader middlewares
 # See https://docs.scrapy.org/en/latest/topics/downloader-middleware.html
 DOWNLOADER_MIDDLEWARES = {
-   'scrapy_po.InjectionMiddleware': 543,
+   'scrapy_poet.InjectionMiddleware': 543,
 }
 
 # Enable or disable extensions
