+++ example/example/spiders/books_02_2.py
@@ -13,7 +13,7 @@ it is better than defining callback explicitly.
 """
 import scrapy
 from web_poet.pages import WebPage
-from scrapy_po import callback_for
+from scrapy_poet import callback_for
 
 
 class BookPage(WebPage):
