+++ example/example/spiders/books_04.py
@@ -4,7 +4,7 @@ Scrapy spider which uses Page Objects both for crawling and extraction.
 """
 import scrapy
 from web_poet.pages import WebPage
-from scrapy_po import callback_for
+from scrapy_poet import callback_for
 
 
 class BookListPage(WebPage):
