+++ example/example/spiders/books_05_1.py
@@ -15,7 +15,7 @@ and the ``BooksSpider.parse_book`` callback for implementation details.
 import scrapy
 
 from web_poet.pages import WebPage
-from scrapy_po.utils import DummyResponse
+from scrapy_poet.utils import DummyResponse
 from example.autoextract import ProductPage
