+++ scrapy_poet/utils.py
@@ -8,10 +8,10 @@ from scrapy.utils.defer import maybeDeferred_coro
 from twisted.internet.defer import inlineCallbacks, returnValue
 
 from web_poet.pages import ItemPage, is_injectable
-from scrapy_po.page_input_providers import providers
+from scrapy_poet.page_input_providers import providers
 
 
-_CALLBACK_FOR_MARKER = '__scrapy_po_callback'
+_CALLBACK_FOR_MARKER = '__scrapy_poet_callback'
 
 
 def get_callback(request, spider):
