+++ tox.ini
@@ -11,12 +11,12 @@ deps =
 
 commands =
     py.test \
-        --cov-report=term --cov-report=html --cov-report= --cov=scrapy_po \
+        --cov-report=term --cov-report=html --cov-report= --cov=scrapy_poet \
         --doctest-modules \
-        {posargs:scrapy_po tests}
+        {posargs:scrapy_poet tests}
 
 [testenv:mypy]
 deps =
     mypy==0.720
 
-commands = mypy --ignore-missing-imports --no-warn-no-return scrapy_po tests
+commands = mypy --ignore-missing-imports --no-warn-no-return scrapy_poet tests
